+++ CHANGELOG.md
@@ -5,3 +5,10 @@ Release as a standalone library (with android analytics support).
 ## 1.0.1
 
 Added getSkarbDeviceId method.
+
+## 2.0.0
+
+Added SkarbUserPurchaseInfo with info about the user's purchase history
+Returning SkarbUserPurchaseInfo from fetchUserPurchasesInfo method
+**fetchUserPurchasesInfo method may now throw an exception!**
+Reworked purchasePackage return type